requirements.txt
google-cloud-aiplatform==1.38.0
